Refactored SequencingExperimentConverter.
