+++ api/ontology.py
@@ -44,3 +44,6 @@ class OntologyAPI:
 
 class Error(Exception):
     """Base-class for all exceptions raised by this module."""
+
+
+__api__ = OntologyAPI()
